@@ -13,10 +13,10 @@
     "@soerenmartius/vue3-clipboard": "^0.1.2",
     "@vue/compat": "3.4.15",
     "axios": "^1.1.3",
-    "bootstrap": "^5.1.3",
-    "bootstrap-vue-3": "^0.5.1",
+    "bootstrap": "^5.3.2",
+    "bootstrap-vue-next": "^0.23.5",
     "core-js": "^3.21.1",
-    "filesize": "^10.0.12",
+    "filesize": "10.1.4",
     "json2yaml": "^1.1.0",
     "moment": "^2.24.0",
     "vue": "3.4.15",
@@ -30,15 +30,18 @@
     "@vue/cli-plugin-eslint": "^5.0.4",
     "@vue/cli-plugin-unit-jest": "~5.0.0",
     "@vue/cli-service": "^5.0.4",
-    "@vue/compiler-sfc": "^3.2.31",
+    "@vue/compiler-sfc": "^3.4.15",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/test-utils": "^2.0.0-0",
     "@vue/vue3-jest": "^27.0.0-alpha.1",
     "babel-jest": "^27.0.6",
-    "eslint": "^8.11.0",
+    "deepmerge": "^4.3.1",
+    "eslint": "^9.6.0",
     "eslint-plugin-vue": "^9.6.0",
     "flush-promises": "^1.0.2",
     "jest": "^27.0.5",
+    "sass": "^1.71.1",
+    "sass-loader": "^14.1.1",
     "webpack": "^5.7.0"
   }
 }
