@@ -1,21 +1,20 @@
-import { createApp } from "vue";
-import App from "./App.vue";
-import { store } from './store'
+import { createApp } from 'vue';
+import App from './App.vue';
+import { store } from './store';
 
-import axios from "axios";
-import BootstrapVue3 from 'bootstrap-vue-3';
-import VueAxios from 'vue-axios'
+import axios from 'axios';
+import { createBootstrap } from 'bootstrap-vue-next'; 
+import VueAxios from 'vue-axios';
 import { VueClipboard } from '@soerenmartius/vue3-clipboard';
 
-import "bootstrap/dist/css/bootstrap.css";
-import "bootstrap-vue-3/dist/bootstrap-vue-3.css";
+import './assets/css/custom.scss';
+import 'bootstrap-vue-next/dist/bootstrap-vue-next.css'; 
 
 const app = createApp(App);
 
-
-app.use(BootstrapVue3);
+app.use(createBootstrap());
 app.use(VueClipboard);
 app.use(store);
-app.use(VueAxios, axios)
+app.use(VueAxios, axios);
 app.config.productionTip = false;
-app.mount("#app");
+app.mount('#app');
